Fixed some warnings

+++ src/utility/FastMath.cpp
@@ -9,7 +9,7 @@
 float FastMath::fastMean(int startPosition, unsigned int nItems, const std::vector<int> *values) {
     // Compute the mean of the first nItems elements of the vector
     int mean = 0;
-    for (auto i = startPosition; i < startPosition + nItems; i++) {
+    for (auto i = startPosition; i < startPosition + (int) nItems; i++) {
         int v = values->at(i);
         mean += v;
     }
@@ -20,7 +20,7 @@ float FastMath::fastStdDev(int startPosition, unsigned int nItems, const std::ve
     // Compute the standard deviation of the first nItems elements of the vector
     float mean = fastMean(startPosition, nItems, values);
     float sumSqDiff = 0;
-    for (auto i = startPosition; i < startPosition + nItems; i++) {
+    for (auto i = startPosition; i < startPosition + (int) nItems; i++) {
         int v = values->at(i);
         float diff = static_cast<float>(v) - mean;
         sumSqDiff += diff * diff;
